+++ packages/medusa/src/api/routes/admin/orders/fulfill-swap.ts
@@ -37,7 +37,9 @@ import { updateInventoryAndReservations } from "./create-fulfillment"
  *       import Medusa from "@medusajs/medusa-js"
  *       const medusa = new Medusa({ baseUrl: MEDUSA_BACKEND_URL, maxRetries: 3 })
  *       // must be previously logged in or use api token
- *       medusa.admin.orders.fulfillSwap(order_id, swap_id)
+ *       medusa.admin.orders.fulfillSwap(order_id, swap_id, {
+ *         
+ *       })
  *       .then(({ order }) => {
  *         console.log(order.id);
  *       });